\name{probLinkMetropolis-package}
\alias{probLinkMetropolis-package}
\alias{probLinkMetropolis}
\docType{package}
\title{
\packageTitle{probLinkMetropolis}
}
\description{
\packageDescription{probLinkMetropolis}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{probLinkMetropolis}
\packageIndices{probLinkMetropolis}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{probLinkMetropolis}

Maintainer: \packageMaintainer{probLinkMetropolis}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
